requirements.txt
torch
transformers
fastapi
uvicorn
pillow
